requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
mangum==0.17.0
boto3==1.34.144
pydantic==2.7.4
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
